보드 초기화, 보드판 그림 그려주기, 방향키 , 게임 끝내기 분할 컴파일 완료
