+++ ScoreAndScheme.h
@@ -1,0 +1,10 @@
+#ifndef _SCOREANDSCHEME_H
+
+#define _SCOREANDSCHEME_H
+
+#include <stdint.h>
+
+uint32_t score;	//unsigned는 부호비트를 제거해서 양수 범위를 2배로 늘리기 위해 사용
+uint8_t scheme;
+
+#endif
