+++ Size.h
@@ -1,0 +1,11 @@
+#ifndef _SIZE_H
+
+#define _SIZE_H
+
+#include <stdint.h>
+
+#define SIZE 4
+
+uint8_t board[SIZE][SIZE];
+
+#endif
