+++ initBoard.h
@@ -1,0 +1,11 @@
+#ifndef _INITBOARD_H
+
+#define _INITBOARD_H
+
+#include"Size.h"
+#include"ScoreAndScheme.h"
+#include"DrawBoard.h"
+
+void initBoard(uint8_t board[SIZE][SIZE]);
+
+#endif
