--- testheader.h
@@ -1,9 +1,0 @@
-#define _XOPEN_SOURCE 500
-
-#include <stdint.h>
-
-
-#define SIZE 4
-uint32_t score = 0;	//unsigned는 부호비트를 제거해서 양수 범위를 2배로 늘리기 위해 사용
-uint8_t scheme = 0;
-
